+++ app/src/main/java/com/agraeta/user/btl/Shopping_Product_view.java
@@ -809,8 +809,8 @@ public class Shopping_Product_view extends AppCompatActivity {
                             //double maxqu = Double.parseDouble(maxq);
                             int qu = Integer.parseInt(qty);
 
-                            double a = qu / buyqu;
-                            double b = a * getqu;
+                            int a = (int) (qu / buyqu);
+                            int b = (int) (a * getqu);
                             /*if (b > maxqu) {
                                 b = maxqu;
                             } else {
@@ -828,7 +828,7 @@ public class Shopping_Product_view extends AppCompatActivity {
                                 jobject.put("category_id", bean_product_schme.get(0).getPro_cat_id());
                                 jobject.put("name", bean_product_schme.get(0).getPro_name());
                                 jobject.put("pro_code", bean_product_schme.get(0).getPro_code());
-                                jobject.put("quantity", String.valueOf((int) b));
+                                jobject.put("quantity", String.valueOf(b));
                                 jobject.put("mrp", bean_product_schme.get(0).getPro_mrp());
                                 //Log.e("D1D1",""+bean_product_schme.get(0).getPro_mrp());
                                 jobject.put("selling_price", bean_product_schme.get(0).getPro_sellingprice());
@@ -3258,8 +3258,8 @@ public class Shopping_Product_view extends AppCompatActivity {
                                             //double maxqu = Double.parseDouble(maxq);
                                             int qu = Integer.parseInt(edt_count.getText().toString());
 
-                                            double a = qu / buyqu;
-                                            double b = a * getqu;
+                                            int a = (int) (qu / buyqu);
+                                            int b = (int) (a * getqu);
                                            /* if (b > maxqu) {
                                                 b = maxqu;
                                             } else {
@@ -3267,7 +3267,7 @@ public class Shopping_Product_view extends AppCompatActivity {
                                             }*/
 
 
-                                            bean_s.setPro_qty(String.valueOf((int) b));
+                                            bean_s.setPro_qty(String.valueOf(b));
 
 
                                             bean_s.setPro_mrp("0");
@@ -3304,7 +3304,7 @@ public class Shopping_Product_view extends AppCompatActivity {
                                                 jobject.put("category_id", bean_product_schme.get(0).getPro_cat_id());
                                                 jobject.put("name", bean_product_schme.get(0).getPro_name());
                                                 jobject.put("pro_code", bean_product_schme.get(0).getPro_code());
-                                                jobject.put("quantity", String.valueOf((int) b));
+                                                jobject.put("quantity", String.valueOf(b));
                                                 jobject.put("mrp", bean_product_schme.get(0).getPro_mrp());
                                                 jobject.put("selling_price", bean_product_schme.get(0).getPro_sellingprice());
                                                 jobject.put("option_id", bean_Oprtions.get(0).getPro_Option_id().toString());
@@ -6379,8 +6379,8 @@ public class Shopping_Product_view extends AppCompatActivity {
                                         //double maxqu = Double.parseDouble(maxq);
                                         int qu = Integer.parseInt(qty);
 
-                                        double a = qu / buyqu;
-                                        double b = a * getqu;
+                                        int a = (int) (qu / buyqu);
+                                        int b = (int) (a * getqu);
                                         /*if (b > maxqu) {
                                             b = maxqu;
                                         } else {
@@ -6398,7 +6398,7 @@ public class Shopping_Product_view extends AppCompatActivity {
                                             jobject.put("category_id", bean_product_schme.get(0).getPro_cat_id());
                                             jobject.put("name", bean_product_schme.get(0).getPro_name());
                                             jobject.put("pro_code", bean_product_schme.get(0).getPro_code());
-                                            jobject.put("quantity", String.valueOf((int) b));
+                                            jobject.put("quantity", String.valueOf(b));
                                             jobject.put("mrp", bean_product_schme.get(0).getPro_mrp());
                                             //Log.e("D1D1",""+bean_product_schme.get(0).getPro_mrp());
                                             jobject.put("selling_price", bean_product_schme.get(0).getPro_sellingprice());
